requirements.txt
fastapi
uvicorn
python-multipart
androguard
thefuzz
pillow
jinja2
reportlab
torch
transformers
imagehash
